Fix migration and seeder
